+++ Sources/Resources/Device.swift
@@ -6,7 +6,7 @@
 import Foundation
 
 /// A device registered in your SimpleMDM account.
-public struct Device: ListableResource {
+public struct Device: ListableResource, SearchableResource {
     // sourcery:inline:auto:Device.Identifiable
     /// The unique identifier of this resource.
     public let id: Int