+++ Tests/NetworkingTests.swift
@@ -206,6 +206,30 @@ internal class NetworkingTests: XCTestCase {
         }
     }
 
+    func testMalformedResourceSearchURL() {
+        struct FakeResource: SearchableResource {
+            typealias Identifier = String
+
+            var id: String
+            static var endpointName: String { return "💩" }
+        }
+
+        let session = URLSessionMock()
+        let s = SimpleMDM(sessionMock: session)
+
+        let cursor = SearchCursor<FakeResource>(searchString: "Foobar")
+        cursor.next(s.networking) { result in
+            guard case let .failure(error) = result else {
+                return XCTFail("Expected .failure, got \(result)")
+            }
+            guard let internalError = error as? InternalError else {
+                return XCTFail("Expected error to be a InternalError, got \(error)")
+            }
+            XCTAssertEqual(internalError, InternalError.malformedURL)
+            XCTAssertEqual(internalError.localizedDescription, "The URL could not be constructed")
+        }
+    }
+
     func testMalformedNestedResourceURL() {
         struct FakeResource: IdentifiableResource {
             typealias Identifier = String